+++ packages/graphback-core/src/runtime/interfaces.ts
@@ -14,9 +14,7 @@ export interface GraphbackServiceConfigMap {
  * GraphQL context interface according to Graphback runtime layer format
  */
 export interface GraphbackContext {
-  graphback: {
-    services: GraphbackServiceConfigMap
-  }
+  graphback: GraphbackServiceConfigMap
 }
 
 /**